[refactor] (litecoin) move common HD test to helper.ts
